huge refactoring, split code into several submodules
