+++ testgres/__init__.py
@@ -1,1 +1,20 @@
-from .testgres import *
+from .api import get_new_node
+from .backup import NodeBackup
+from .config import TestgresConfig, configure_testgres
+
+from .connection import \
+    IsolationLevel, \
+    NodeConnection, \
+    InternalError, \
+    ProgrammingError
+
+from .exceptions import *
+from .node import NodeStatus, PostgresNode
+
+from .utils import \
+    reserve_port, \
+    release_port, \
+    bound_ports, \
+    get_bin_path, \
+    get_pg_config, \
+    get_pg_version
