+++ testgres/consts.py
@@ -1,0 +1,11 @@
+# names for dirs in base_dir
+DATA_DIR = "data"
+LOGS_DIR = "logs"
+
+# names for log files
+PG_LOG_FILE = "postgresql.log"
+UTILS_LOG_FILE = "utils.log"
+BACKUP_LOG_FILE = "backup.log"
+
+# default argument value
+DEFAULT_XLOG_METHOD = "fetch"
